Move SCFToCFG into Util dialect and later in the pipeline
It is difficult to investigate buffer / dispatch behavior across
CFG, while SCF is much simpler to trace their usage. Pushed SCF
to CFG later in the pipeline.
