+++ compiler/src/iree/compiler/Dialect/Util/Transforms/BUILD.bazel
@@ -36,6 +36,7 @@ iree_compiler_cc_library(
         "StripDebugOps.cpp",
         "TestConversion.cpp",
         "TestFloatRangeAnalysis.cpp",
+        "TopLevelSCFToCFG.cpp",
     ],
     hdrs = [
         "Passes.h",
