+++ compiler/src/iree/compiler/Dialect/Util/Transforms/CMakeLists.txt
@@ -39,6 +39,7 @@ iree_cc_library(
     "StripDebugOps.cpp"
     "TestConversion.cpp"
     "TestFloatRangeAnalysis.cpp"
+    "TopLevelSCFToCFG.cpp"
   DEPS
     ::PassesIncGen
     LLVMSupport
