+++ compiler/src/iree/compiler/Dialect/Util/Transforms/Passes.h
@@ -8,6 +8,7 @@
 #define IREE_COMPILER_DIALECT_IREE_TRANSFORMS_PASSES_H_
 
 #include "mlir/IR/BuiltinOps.h"
+#include "mlir/IR/FunctionInterfaces.h"
 #include "mlir/Pass/Pass.h"
 #include "mlir/Pass/PassManager.h"
 
@@ -31,6 +32,7 @@ std::unique_ptr<OperationPass<void>> createSimplifyGlobalAccessesPass();
 std::unique_ptr<OperationPass<mlir::ModuleOp>>
 createStripAndSplatConstantsPass();
 std::unique_ptr<OperationPass<void>> createStripDebugOpsPass();
+std::unique_ptr<InterfacePass<mlir::FunctionOpInterface>> createTopLevelSCFToCFGPass();
 
 // Resource Management.
 std::unique_ptr<OperationPass<void>> createImportResourcesPass();
