+++ compiler/src/iree/compiler/Dialect/Util/Transforms/test/CMakeLists.txt
@@ -36,6 +36,7 @@ iree_lit_test_suite(
     "strip_debug_ops.mlir"
     "test_float_range_analysis.mlir"
     "test_float_range_analysis_linalg.mlir"
+    "top_level_scf_to_cfg.mlir"
   TOOLS
     FileCheck
     iree-opt
