+++ compiler/src/iree/compiler/InputConversion/Common/BUILD.bazel
@@ -50,7 +50,6 @@ iree_compiler_cc_library(
         "QuantizedConvToConv.cpp",
         "QuantizedMatmulToMatmul.cpp",
         "SanitizeModuleNames.cpp",
-        "TopLevelSCFToCFG.cpp",
         "Utils.cpp",
     ],
     hdrs = [
