+++ compiler/src/iree/compiler/InputConversion/Common/Passes.h
@@ -41,7 +41,6 @@ createLinalgQuantizedConvToConvPass();
 std::unique_ptr<OperationPass<func::FuncOp>>
 createLinalgQuantizedMatmulToMatmulPass();
 std::unique_ptr<OperationPass<ModuleOp>> createSanitizeModuleNamesPass();
-std::unique_ptr<OperationPass<func::FuncOp>> createTopLevelSCFToCFGPass();
 
 //===----------------------------------------------------------------------===//
 // Register all Passes
