+++ compiler/src/iree/compiler/InputConversion/Common/test/CMakeLists.txt
@@ -20,7 +20,6 @@ iree_lit_test_suite(
     "linalg_quantized_conv_to_conv.mlir"
     "linalg_quantized_matmul_to_matmul.mlir"
     "sanitize_module_names.mlir"
-    "top_level_scf_to_cfg.mlir"
   TOOLS
     FileCheck
     iree-opt
